synthetic_hate/requirements.txt
groq>=0.4.0
transformers>=4.38.0
datasets>=2.18.0
scikit-learn>=1.4.0
sentence-transformers>=2.6.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
tqdm>=4.66.0
torch>=2.2.0
evaluate>=0.4.1
accelerate>=0.27.0
joblib>=1.3.0
